refactor NewPerpetualForcedActionPage once again
